Add files via upload
This is a childrens game I made during school for a final project. I used my daughter for inspiration as she was 1 when I started making this. It has some issues like one feature is that is has an automated speech function when you click the button. If you spam the button the voice overlays and starts to get very loud and annoying. The other issue is that there is no home function until you finish the game. This is by design so that the child cannot leave the intended game until finished.